Show very simple version of artist card
The ArtistsCard does not show any images anymore. We cannot show any
Roon images for two reasons. First, we want to avoid the Roon API call
to get the images. And second, the MusicBrainz API may return multiple
artists although we start with a single artist name that we obtain
when asking the Roon API for album data.